+++ src/classes/scripts.ts
@@ -46,7 +46,7 @@ export class Scripts {
     queue: MinimalQueue,
     listKey: string,
     jobId: string,
-  ) {
+  ): Promise<boolean> {
     const client = await queue.client;
     let result;
     if (semver.lt(queue.redisVersion, '6.0.6')) {
@@ -259,7 +259,7 @@ export class Scripts {
     }
   }
 
-  static drainArgs(queue: MinimalQueue, delayed: boolean) {
+  static drainArgs(queue: MinimalQueue, delayed: boolean): string[] {
     const queueKeys = queue.keys;
 
     const keys = [
@@ -274,11 +274,11 @@ export class Scripts {
     return keys.concat(args);
   }
 
-  static async drain(queue: MinimalQueue, delayed: boolean) {
+  static async drain(queue: MinimalQueue, delayed: boolean): Promise<void> {
     const client = await queue.client;
     const args = this.drainArgs(queue, delayed);
 
-    await (<any>client).drain(args);
+    return (<any>client).drain(args);
   }
 
   static moveToCompleted(
@@ -365,7 +365,8 @@ export class Scripts {
   ): Promise<void> {
     const client = await queue.client;
 
-    const args = this.changeDelayArgs(queue, jobId, delay);
+    const delayTimestamp = Date.now() + delay;
+    const args = this.changeDelayArgs(queue, jobId, delayTimestamp);
     const result = await (<any>client).changeDelay(args);
     if (result < 0) {
       throw this.finishedErrors(result, jobId, 'changeDelay', 'delayed');